+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 - Added option to start application even if Java version requirements are not met.
 - Changed accelerator of "Copy Throwable" from "command shift alt T" to "command shift T".
 - Changed specification of HTTP status code 451 from draft to RFC 7725.
+- Changed meaning of `ThrowableCondition`. Condition now evaluates to true if search string is one of the `Throwable` contained in the hierarchy (including cause, suppressed) instead of only checking the root exception.
 
 ### Deprecated
 - Nothing.